Moved StickyToolbar component to toolbar/sticky/.
